bugfix: stringify search query object
When querying API with search query ('q' query param), in the hood
superagent is using native JS function 'encodeURIComponent' to serialize
each query param, changing complex Objects to '[object Object]'.
To not force clients to reimplement 'SearchQuery.property.toString' we are
stringifying it before it gets to superagent, making this truly transparent.

Relates to kisi-inc/web-react#149.