Fix for double slash issue
If someone will pass to the path value with leading slash e.g. `/abc` we will concatenate it as a `https://hostname//abc`, which is not correct.

This changeset resolves that issue, and makes that more *bulletproof* for the future. :wink:
I have bumped up patch version, because it should not affect anyone (it is actually kind of *bugfix*).